Fix Windows Auto-Type sending characters to virtualized guests
* Fix #1833

+++ src/gui/osutils/winutils/WinUtils.h
@@ -54,7 +54,7 @@ public:
                                 QString* error = nullptr) override;
     bool unregisterGlobalShortcut(const QString& name) override;
 
-    DWORD qtToNativeKeyCode(Qt::Key key);
+    WORD qtToNativeKeyCode(Qt::Key key);
     DWORD qtToNativeModifiers(Qt::KeyboardModifiers modifiers);
 
     bool canPreventScreenCapture() const override;
